+++ product-service/src/main/java/com/creditApp/model/dto/ProductDto.java
@@ -7,6 +7,7 @@ import lombok.Setter;
 
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -16,6 +17,6 @@ public class ProductDto {
     @NotEmpty(message = "{product.name.notEmpty}")
     private String name;
     @NotNull(message = "{product.value.notNull}")
-    private double value;
+    private BigDecimal value;
     private String creditNumber;
 }
